Create new file structure, split up actions.jsx
