+++ client/src/actions/cartActions.jsx
@@ -1,6 +1,6 @@
-export const ADD_TO_CART = 'ADD_TO_CART';
+// export const ADD_TO_CART = 'ADD_TO_CART';
 
-export const addToCart = product => ({
-  type: ADD_TO_CART,
-  product,
-});
+// export const addToCart = product => ({
+//   type: ADD_TO_CART,
+//   product,
+// });
