--- client/src/components/Link.jsx
@@ -1,40 +1,0 @@
-import React from 'react';
-
-const STATUS = {
-  HOVERED: 'testing',
-  NORMAL: 'normal',
-};
-
-export default class Link extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this._onMouseEnter = this._onMouseEnter.bind(this);
-    this._onMouseLeave = this._onMouseLeave.bind(this);
-
-    this.state = {
-      class: [STATUS.NORMAL, 'randomly'],
-    };
-  }
-
-  _onMouseEnter() {
-    this.setState({ class: STATUS.HOVERED });
-  }
-
-  _onMouseLeave() {
-    this.setState({ class: STATUS.NORMAL });
-  }
-
-  render() {
-    return (
-      <a
-        className={this.state.class}
-        href={this.props.page || '#'}
-        onMouseEnter={this._onMouseEnter}
-        onMouseLeave={this._onMouseLeave}
-      >
-        {this.props.children}
-      </a>
-    );
-  }
-}
