--- client/src/components/__tests__/Link.test.js
@@ -1,24 +1,0 @@
-// Link.react-test.js
-import React from 'react';
-import Link from '../Link.jsx';
-import renderer from 'react-test-renderer';
-
-test('Link changes the class when hovered', () => {
-  const component = renderer.create(
-    <Link page="http://www.facebook.com">Facebook</Link>,
-  );
-  let tree = component.toJSON();
-  expect(tree).toMatchSnapshot();
-
-  // manually trigger the callback
-  tree.props.onMouseEnter();
-  // re-rendering
-  tree = component.toJSON();
-  expect(tree).toMatchSnapshot();
-
-  // manually trigger the callback
-  tree.props.onMouseLeave();
-  // re-rendering
-  tree = component.toJSON();
-  expect(tree).toMatchSnapshot();
-});
