+++ client/src/components/customer/Cart.jsx
@@ -1,0 +1,10 @@
+// import React from 'react';
+
+// const Cart = (props) => {
+//   return (
+//     <div>
+//     </div>
+//   )
+// }
+
+// export default Cart;
