+++ client/src/reducers/cartReducer.jsx
@@ -1,14 +1,14 @@
-import {
-  ADD_TO_CART,
-} from '../actions/cartActions.jsx';
+// import {
+//   ADD_TO_CART,
+// } from '../actions/cartActions.jsx';
 
-export default function cart(state = [], action) {
-  switch (action.type) {
-    case ADD_TO_CART: {
-      return [...state, action.product];
-    }
-    default: {
-      return state;
-    }
-  }
-}
+// export default function cart(state = [], action) {
+//   switch (action.type) {
+//     case ADD_TO_CART: {
+//       return [...state, action.product];
+//     }
+//     default: {
+//       return state;
+//     }
+//   }
+// }
